notes: Fix error when saving a note with media content w/out caption in some cases
If media(documents, photos, audio, etc...) is saved without a caption, the save command will use the note name as the caption. If the note name has underscores or other Markdown characters, Telegram will try parsing them. If there is an odd number of them, it will throw a Bad Request error. This change will escape markdown in those cases.

Signed-off-by: Gegham Zakaryan <zakaryan.2004@outlook.com>